Change amount to int

+++ PaymentsSharing/Payments/AddPayment.cs
@@ -6,6 +6,6 @@ namespace PaymentsSharing.Payments;
 internal record AddPayment(
     IReadOnlyCollection<Person> Payers,
     IReadOnlyCollection<Person> Consumers,
-    decimal Amount,
-    decimal? AmountForMeat = null,
+    uint Amount,
+    uint? AmountForMeat = null,
     string Description = "") : IRequest;
